replace Doubles with Floats/float[] when possible (MongoDB works with Doubles internally)

+++ langchain4j-mongodb/src/main/java/dev/langchain4j/store/embedding/mongodb/EmbeddingDocument.java
@@ -19,7 +19,7 @@ public class EmbeddingDocument {
     @NonNull
     private String id;
     @NonNull
-    private List<Double> embedding;
+    private List<Float> embedding;
 
     private String text;
     private Map<String, String> metadata;
